refactor: :zap: Replaced double prices with integers representing number of cents
Replaced double prices with ints that store the number of cents (142.50 now 14250); Eliminates floating point arithmetic and extra CPU cycles on the hot path; Matching engine reflects all needed changes; Post-enum optimization (650 ns) brought down to 515 ns average with 4 ns variance

+++ order-book/OrderBook.hpp
@@ -19,8 +19,8 @@ namespace Order {
             void reset();
     
             // Market info methods
-            double getSpread();
-            double getLastTradedPrice();
+            int getSpread();
+            int getLastTradedPrice();
             int getSize();
     
         private:
@@ -82,7 +82,7 @@ namespace Order {
             void removeAsk(int heapLoc);
             void removeBid(int heapLoc);
 
-            double lastTradedPrice;
+            int lastTradedPrice;
     };    
 }
